+++ lib/sync_datetime.dart
@@ -1,4 +1,5 @@
 /// Support for reliable and predictable DateTime synchronization.
 library;
 
+export 'src/models/server_datetime_parts_model.dart';
 export 'src/sync_datetime.dart';